requirements.txt
networkx~=2.8.8
numpy~=1.21.5
Django==2.2.27
pandas~=1.4.2
PyYAML~=6.0
concurrent_log==1.0.1
requests~=2.27.1
joblib~=1.1.0
gunicorn==20.1.0
pymysql
psycopg2
taospy==2.6.4
matplotlib==3.5.1